# Refactor web.NewLine to use reflection. Increases performance by ~240%