- Replace json.Marshal/Unmarshal with reflection to determine how to write
  data back to user
- Increase performance by about 240%

Previous Benchmark:
```
    BenchmarkNewLine-8          3000            576894 ns/op
    BenchmarkNewLine-8          2000            624219 ns/op
    BenchmarkNewLine-8          3000            666386 ns/op

    (avg ~622000 ns/op)
```

New Benchmark:
```
    BenchmarkNewLine-8          5000            244293 ns/op
    BenchmarkNewLine-8          5000            265308 ns/op
    BenchmarkNewLine-8          5000            247538 ns/op

    (avg 252000 ns/op), ~2.42x speedup
```